Generate code to lazily instantiate templates
It is a bit ugly (maybe we should add a `memoize(Supplier)` utility somewhere), but the returned visitor now no longer eagerly instantiates the `JavaTemplate.Builder` and thus will often avoid the cost of classpath scanning. This is especially important for the case when the visitor won't even run due to the preconditions.